+++ PurchaseManagement/MVVM/Models/MarketModels/Product.cs
@@ -3,14 +3,14 @@
     public class Product:BaseEntity
     {
         public string Item_Name { get; set; }
-        public long Item_Price { get; set; }
-        public long Item_Quantity { get; set; }
+        public double Item_Price { get; set; }
+        public double Item_Quantity { get; set; }
         public string Item_Description { get; set; }
         public bool IsPurchased { get; set; }
         public virtual ProductLocation ProductLocation { get; set; }
         public int PurchaseId { get; set; }
         public virtual Purchase Purchase { get; set; }
-        public Product(string item_name, long item_price, long item_quantity, string item_desc)
+        public Product(string item_name, double item_price, double item_quantity, string item_desc)
         {
             Item_Name = item_name;
             Item_Price = item_price;
